pallet-cosmwasm: reduce MaxFrames parameter to u16
65k recursive contract calls should be enough for anything so there’s
no real need to use MaxFrames parameter of larger type than u16.  With
that, we can replace a somewhat noisy usize::try_from call with a more
straightforward usize::from.

+++ code/parachain/frame/cosmwasm/src/lib.rs
@@ -230,7 +230,7 @@ pub mod pallet {
 
 		/// Max number of frames a contract is able to push, a.k.a recursive calls.
 		#[pallet::constant]
-		type MaxFrames: Get<u32>;
+		type MaxFrames: Get<u16>;
 
 		/// Max accepted code size in bytes.
 		#[pallet::constant]
